Refactor packages (#625)
