+++ cassandra/src/test/scala/com/stratio/crossdata/connector/cassandra/CassandraQueryProcessorSpec.scala
@@ -17,7 +17,7 @@ package com.stratio.crossdata.connector.cassandra
 
 import com.stratio.crossdata.test.BaseXDTest
 import org.apache.spark.sql.catalyst.expressions.{AttributeReference, Literal}
-import org.apache.spark.sql.crossdata.execution.NativeUDF
+import org.apache.spark.sql.crossdata.catalyst.NativeUDF
 import org.apache.spark.sql.sources
 import org.apache.spark.sql.types.DataTypes
 import org.junit.runner.RunWith
