+++ core/src/main/scala/com/stratio/crossdata/connector/SQLLikeUDFQueryProcessorUtils.scala
@@ -20,7 +20,7 @@ import java.sql.Timestamp
 import org.apache.spark.sql.catalyst.expressions.Attribute
 import org.apache.spark.sql.catalyst.expressions.AttributeReference
 import org.apache.spark.sql.catalyst.expressions.Literal
-import org.apache.spark.sql.crossdata.execution.NativeUDF
+import org.apache.spark.sql.crossdata.catalyst.NativeUDF
 import org.apache.spark.sql.types.DataTypes
 
 object SQLLikeUDFQueryProcessorUtils {
