+++ core/src/main/scala/org/apache/spark/sql/crossdata/catalyst/planning/XDStrategies.scala
@@ -13,14 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.spark.sql.crossdata.execution
+package org.apache.spark.sql.crossdata.catalyst.planning
 
 import org.apache.spark.sql.Strategy
-import org.apache.spark.sql.catalyst.TableIdentifier
 import org.apache.spark.sql.catalyst.plans.logical.LogicalPlan
 import org.apache.spark.sql.crossdata.XDContext
 import org.apache.spark.sql.crossdata.catalog.TableIdentifierNormalized
 import org.apache.spark.sql.crossdata.catalog.XDCatalog.CrossdataTable
+import org.apache.spark.sql.crossdata.catalyst.execution.{PersistDataSourceTable, PersistSelectAsTable, RegisterDataSourceTable}
 import org.apache.spark.sql.execution.datasources.{CreateTableUsing, CreateTableUsingAsSelect}
 import org.apache.spark.sql.execution.{ExecutedCommand, SparkPlan, SparkStrategies}
