+++ core/src/test/scala/org/apache/spark/sql/crossdata/XDContextIT.scala
@@ -18,7 +18,7 @@ package org.apache.spark.sql.crossdata
 import java.nio.file.Paths
 
 import org.apache.spark.sql.catalyst.TableIdentifier
-import org.apache.spark.sql.crossdata.execution.PersistDataSourceTable
+import org.apache.spark.sql.crossdata.catalyst.execution.PersistDataSourceTable
 import org.apache.spark.sql.crossdata.test.SharedXDContextTest
 import org.apache.spark.sql.execution.ExecutedCommand
 import org.apache.spark.sql.types.{IntegerType, StringType, StructField, StructType}
