+++ core/src/test/scala/org/apache/spark/sql/crossdata/XDSessionIT.scala
@@ -23,7 +23,7 @@ import com.typesafe.config.ConfigFactory
 import org.apache.spark.sql.catalyst.TableIdentifier
 import org.apache.spark.sql.crossdata.catalog.persistent.DerbyCatalog
 import org.apache.spark.sql.crossdata.catalog.temporary.HashmapCatalog
-import org.apache.spark.sql.crossdata.execution.PersistDataSourceTable
+import org.apache.spark.sql.crossdata.catalyst.execution.PersistDataSourceTable
 import org.apache.spark.sql.crossdata.session.{XDSessionState, XDSharedState}
 import org.apache.spark.sql.execution.ExecutedCommand
 import org.apache.spark.sql.types.{IntegerType, StringType, StructField, StructType}
