+++ testsIT/src/test/scala/com/stratio/crossdata/driver/globalindex/InsertGlobalIndexIT.scala
@@ -16,7 +16,7 @@
 package com.stratio.crossdata.driver.globalindex
 
 import com.sksamuel.elastic4s.ElasticDsl._
-import org.apache.spark.sql.crossdata.execution.datasources.CreateGlobalIndex
+import org.apache.spark.sql.crossdata.catalyst.execution.CreateGlobalIndex
 import org.junit.runner.RunWith
 import org.scalatest.junit.JUnitRunner
